+++ examples/faceted_search_with_tweaked_score.rs
@@ -12,7 +12,7 @@ use std::collections::HashSet;
 use yeehaw::collector::TopDocs;
 use yeehaw::query::BooleanQuery;
 use yeehaw::schema::*;
-use yeehaw::{doc, DocId, Index, IndexWriter, Score, SegmentReader};
+use yeehaw::{DocId, Index, IndexWriter, Score, SegmentReader};
 
 fn main() -> yeehaw::Result<()> {
     let mut schema_builder = Schema::builder();
@@ -25,27 +25,29 @@ fn main() -> yeehaw::Result<()> {
 
     let mut index_writer: IndexWriter = index.writer(30_000_000)?;
 
-    index_writer.add_document(doc!(
-        title => "Fried egg",
-        ingredient => Facet::from("/ingredient/egg"),
-        ingredient => Facet::from("/ingredient/oil"),
-    ))?;
-    index_writer.add_document(doc!(
-        title => "Scrambled egg",
-        ingredient => Facet::from("/ingredient/egg"),
-        ingredient => Facet::from("/ingredient/butter"),
-        ingredient => Facet::from("/ingredient/milk"),
-        ingredient => Facet::from("/ingredient/salt"),
-    ))?;
-    index_writer.add_document(doc!(
-        title => "Egg rolls",
-        ingredient => Facet::from("/ingredient/egg"),
-        ingredient => Facet::from("/ingredient/garlic"),
-        ingredient => Facet::from("/ingredient/salt"),
-        ingredient => Facet::from("/ingredient/oil"),
-        ingredient => Facet::from("/ingredient/tortilla-wrap"),
-        ingredient => Facet::from("/ingredient/mushroom"),
-    ))?;
+    let mut doc = TantivyDocument::new();
+    doc.add_text(title, "Fried egg");
+    doc.add_facet(ingredient, Facet::from("/ingredient/egg"));
+    doc.add_facet(ingredient, Facet::from("/ingredient/oil"));
+    index_writer.add_document(doc)?;
+
+    let mut doc = TantivyDocument::new();
+    doc.add_text(title, "Scrambled egg");
+    doc.add_facet(ingredient, Facet::from("/ingredient/egg"));
+    doc.add_facet(ingredient, Facet::from("/ingredient/butter"));
+    doc.add_facet(ingredient, Facet::from("/ingredient/milk"));
+    doc.add_facet(ingredient, Facet::from("/ingredient/salt"));
+    index_writer.add_document(doc)?;
+
+    let mut doc = TantivyDocument::new();
+    doc.add_text(title, "Egg rolls");
+    doc.add_facet(ingredient, Facet::from("/ingredient/egg"));
+    doc.add_facet(ingredient, Facet::from("/ingredient/garlic"));
+    doc.add_facet(ingredient, Facet::from("/ingredient/salt"));
+    doc.add_facet(ingredient, Facet::from("/ingredient/oil"));
+    doc.add_facet(ingredient, Facet::from("/ingredient/tortilla-wrap"));
+    doc.add_facet(ingredient, Facet::from("/ingredient/mushroom"));
+    index_writer.add_document(doc)?;
     index_writer.commit()?;
 
     let reader = index.reader()?;